Simplifying the Golang/Habitat implementation of license acceptance
The main consumer of this tool is Automate. They run 40+ hab services as
part of the Automate stack, so trying to track each individual service
in the 'product_info.toml' would make their life very difficult. Instead
I have reverted the behavior to be very simple - it simply checks the
presence of the user flag and fails if it is not present. Instead
Automate will manage license acceptance (EULA vs MLSA) in their
application via the browser and 'automate-ctl' tool.

Possible notes for future implementation: maybe we make
`license-acceptance` a dependent hab service instead of making consumers
add an executable to their run script? Currently if the Automate
deployment service calls the license-acceptance binary it must pass all
required configuration to that execution. If it was its own service it
could be configured as an individual service with acceptance for the
entire Automate product, customized license persistence location, etc.
That is basically what they do with their own license service today.

We could also just have 1 product we check - 'Automate'. All the
services ran as part of that could run using this product from
'product_info.toml'.

Signed-off-by: tyler-ball <tball@chef.io>