Adding mixlib-install product search to enable Test Kitchen
Test Kitchen knows products by their mixlib-install name so we add
support for that attribute to the product. We also expose methods to
interact with the products by that attribute. Finally, I added a
convenience method for Test Kitchen to know how the license was accepted
so it can pass that through to provisioned nodes.

Signed-off-by: tyler-ball <tball@chef.io>