Bringing names in line with marketing
This change started by updating the Infra products to `Chef Infra
Client` and `Chef Infra Server`.

In doing this I realized `name` was a bad identifier on the product and
changed it to `id`. This reflects the truth that this is an internal
identifier of a product. It could be a UUID but that would make other
developers jobs hard. `pretty_name` is the name we use when
communicating to users about these products.

Signed-off-by: tyler-ball <tball@chef.io>