Persist both id and pretty_name in the marker files
In the future when we start reading these marker files we can use the
more-permanent 'id' to lookup products. We are more free to change a
product's pretty name and keep that UX focused, while still
communicating to a user what license they have accepted if they open up
the marker file.

Signed-off-by: tyler-ball <tball@chef.io>