EULA as legacy value if not named in license file
Prior to introducing the MLSA acceptance functionality, the EULA was the only possible license
that could be accepted. Prior to this topic branch, the name of the license was not recorded in the license
persistance file. When reading the file, if the license_name field is blank, the only valid value for
it is "EULA", so use that value.

Signed-off-by: Clinton Wolfe <clintoncwolfe@gmail.com>